Attempt to patch issue with SimpleDelegator being undefined with RVM
Including this gem as a dependency works locally, but appears to cause
Codeship to choke due to an uninitialized constant error. Specifically,
the Ruby environment on Codeship seems to be interpreting the reference
to SimpleDelegator as JsonMatchers::SimpleDelegator instead of the
global Rails built-in. Adding an explicit global scope in the hopes it
fixes the issue, though it is unclear why this behaves differently with
RVM/Codeship.